+++ src/Api/Strategies/Strategies.ts
@@ -3,6 +3,7 @@ import { Strategy } from "../Types";
 import checkForSolved from "./checkForSolved";
 import hiddenPairsTriplesAndQuads from "./hiddenPairsTriplesAndQuads";
 import hiddenSingles from "./hiddenSingles";
+import intersectionRemoval from "./intersectionRemoval";
 import pairsTriplesAndQuads from "./pairsTriplesAndQuads";
 import updateCandidates from "./updateCandidates";
 
@@ -12,7 +13,8 @@ const STRATEGIES = [
    updateCandidates,
    hiddenSingles,
    pairsTriplesAndQuads,
-   hiddenPairsTriplesAndQuads
+   hiddenPairsTriplesAndQuads,
+   intersectionRemoval,
 ] as const
 
 export default STRATEGIES as typeof STRATEGIES & Strategy[]